Move Testdata package entityproviding as subpackage of valuerange
